src/requirements.txt
fastapi>=0.110.0
starlette>=0.40.0
uvicorn>=0.27.0
sqlalchemy==2.0.23
psycopg2-binary>=2.9.10
prometheus-client==0.19.0
celery==5.3.4
pika==1.3.2
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-celery==0.42b0
setuptools==78.1.1

# Additional testing dependencies for demos
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For testing FastAPI